billing_system/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
cssselect2==0.7.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
html5lib==1.1
idna==3.10
lxml==5.3.1
oscrypto==1.3.0
pillow==11.1.0
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.3.0
python-bidi==0.6.3
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.1
tzlocal==5.3
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
xhtml2pdf==0.2.16
